:construction: :art: WebViewProvider ahora es compatible con muchos tipos de trabajos simultaneos, Calendario de ETS

+++ app/src/main/java/ziox/ramiro/saes/features/saes/features/schedule/ui/components/ScheduleClassView.kt
@@ -23,7 +23,7 @@ fun ScheduleClassView(
     hourHeight: Dp = 170.dp
 ) = Card(
     modifier = Modifier
-        .padding(top = hourHeight.times((classSchedule.hour.start - startHour).toFloat()))
+        .padding(top = hourHeight.times((classSchedule.hour.start.toDouble() - startHour).toFloat()))
         .height(hourHeight.times(classSchedule.hour.duration.toFloat()))
         .fillMaxWidth(),
     backgroundColor = Color(classSchedule.color),
